work on upward pass
